requirements.txt
aiohttp>=3.9.5
discord.py>=2.3.2
yt-dlp>=2024.10.7
python-dotenv>=1.0.0
PyNaCl>=1.5.0
